fix: resolve compiler warnings in vmm and minip
- Fix format specifier warning in vmm_alloc() by casting to unsigned long
- Add NUSA_INIT_HOOK macro definition to init.h
- Include nusa/init.h in minip.c to resolve macro expansion warning
- Maintain LK_ backward compatibility aliases in init system

Fixes:
* kernel/vm/vmm.c:522 - format '%zu' type mismatch warning
* lib/minip/minip.c:563 - NUSA_INIT_HOOK implicit declaration warning

+++ kernel/vm/vmm.c
@@ -519,7 +519,7 @@ status_t vmm_alloc(vmm_aspace_t *aspace, const char *name, size_t size, void **p
     size_t count = pmm_alloc_pages(size / PAGE_SIZE, &page_list);
     DEBUG_ASSERT(count <= size);
     if (count < size / PAGE_SIZE) {
-        LTRACEF("failed to allocate enough pages (asked for %zu, got %zu)\n", size / PAGE_SIZE, count);
+        LTRACEF("failed to allocate enough pages (asked for %lu, got %lu)\n", (unsigned long)(size / PAGE_SIZE), (unsigned long)count);
         pmm_free(&page_list);
         err = ERR_NO_MEMORY;
         goto err;
